Consolidate cvo and cvo_common; start migration of doutdly methods
