+++ collect/cvo/__init__.py
@@ -4,4 +4,5 @@ collect.cvo
 access cvo data
 """
 # -*- coding: utf-8 -*-
-from . import cvo_common, cvo_dout
+from . import cvo_dout
+from .cvo import *
